refactor(clawdstrike-policy-event): extract kinds/policy_delta from edr/receipt/mod.rs
