+++ crates/libs/clawdstrike-policy-event/src/edr/receipt/kinds/mod.rs
@@ -2,6 +2,7 @@ mod detection;
 mod evidence_bundle;
 mod graph_slice;
 mod policy_decision;
+mod policy_delta;
 mod sensor_state;
 mod simulation;
 
@@ -9,5 +10,6 @@ pub(crate) use detection::*;
 pub(crate) use evidence_bundle::*;
 pub(crate) use graph_slice::*;
 pub(crate) use policy_decision::*;
+pub(crate) use policy_delta::*;
 pub(crate) use sensor_state::*;
 pub(crate) use simulation::*;
